test(postgres): move semicolon-strip tests into postgres job so they actually run
